Add index.html copying and CI Workflow
This update introduces copying of `index.html` to the web directory in `generate.py`. Additionally, a new Python Continuous Integration (CI) workflow has been added under `.github/workflows`. The CI workflow automatically builds the project and deploys it using Cloudflare when changes are pushed or merged to the master branch.