Add welcome alert for first time users
Created a new WelcomeService function to display a welcome alert for first-time users, providing them with options on what information they want to display on their profile. This alert is then called in the TimeService when a project is opened. Also, minor code formatting adjustments have been done and an initialization status has been added in SettingState.