Add option to display default IDE name
Added a configurable checkbox that allows users to choose to display the default JetBrains IDE name instead of the current IDE type. This setting can be found in the theme section and requires a restart to take effect. Also, refactored the applicationId to support this feature and adjusted the initialization of the activity render service accordingly.